refactor(nav): share the floating pill between home and settings
Extract the Home/Settings nav pill into a reusable NavPill and use it on both
the home bottom bar and the settings screen (replacing the settings
NavigationBar), so the two screens present the same floating-pill navigation.
